requirements.txt
imageio==2.8.0
matplotlib==3.1.3
numpy==1.21.2
opencv-python
opencv-contrib-python
Pillow~=8.1.0
tensorflow==2.5.0
tqdm==4.46.0
PyYAML~=5.4.1
albumentations~=1.0.3
torchvision~=0.10.0